% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.pim.R
\docType{methods}
\name{confint.pim}
\alias{confint.pim}
\alias{confint.pim.summary}
\alias{confint}
\alias{confint,pim-method}
\alias{confint,pim.summary-method}
\title{Calculate confidence intervals around the coefficients of a PIM}
\usage{
confint(object, parm, level = 0.95, ...)

\S4method{confint}{pim}(object, parm, level = 0.95, ...)

\S4method{confint}{pim.summary}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{a \code{\link{pim}} or \code{\link{pim.summary}} object}

\item{parm}{a specification of which parameters are to be given confidence intervals. Either a vector of numbers or a vector of names.
If missing, all parameters are considered}

\item{level}{The confidence level required.}

\item{...}{extra arguments to methods}
}
\description{
This function returns confidence intervals around the
coefficients of a fitted \code{\link{pim}} object, assuming asymptotic
normality.
}
\details{
The confidence intervals are calculated using the following formula:
\deqn{
CI = \beta \pm Z_\alpha \frac{\sigma}{\sqrt(n)}
}{
CI = \beta +- Z(\alpha) * \sigma / \sqrt(n)
}
}
\section{Warning}{
 This is a pretty naive way of calculating confidence intervals, and
highly dependent on a correct estimate of the standard error around
the coefficients.
}

\examples{
data('FEVData')
Model <- pim(FEV~ Smoke*Sex , data=FEVData)
confint(Model)
summ <- summary(Model)
confint(summ)
}
