% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.poset.R
\name{create.poset}
\alias{create.poset}
\title{Create a poset}
\usage{
create.poset(compare = c("unique", "all"), n)
}
\arguments{
\item{compare}{a character value, either 'unique' or 'all'}

\item{n}{an single integer value indicating how many observations there
are in the model.}
}
\value{
A named list with 2 elements, called "L" and "R", containing
 the selection indices for the left hand and right hand side of a pim.
}
\description{
This function creates a poset for use in a pim model based on a
number of observations and a comparison type. This function is
called from \code{\link{new.pim.poset}} and returns a list that
can be used as a value for its argument \code{compare}.
}
\examples{
create.poset(n=10)
create.poset('all',n=4)
}
