% Generated by roxygen2 (4.0.1): do not edit by hand
\name{crossvalidate.pim}
\alias{crossvalidate.pim}
\alias{cv.pim}
\alias{cv.pim-class}
\title{Crossvalidate a glmnet PIM}
\usage{
crossvalidate.pim(pimob, method = c("fullsplit", "semisplit", "naive"),
  type.measure = c("mse", "deviance", "class", "auc", "mae"), nfolds = 10,
  foldid, weights, ..., include.extrainfo = FALSE, verbosity = 0)
}
\arguments{
\item{pimob}{Object of class \code{\link{pim}}. Note: it should really be of class
\code{\link{glmnetpim}} in the current implementation.}

\item{method}{\code{"fullsplit"} splits pseudo-observations over folds, \code{"semisplit"} takes
the nonvalidating pseudo-observations together for the fit for each fold, \code{"naive"} simply
crossvalidates over the pseudo-observations (note: typically, this is not correct!)}

\item{type.measure,nfolds,foldid,weights}{See \code{cv.glmnet}}

\item{\dots}{Passed on to \code{cv.glmnet} (or \code{.cvpo.glmnet}).}

\item{include.extrainfo}{If \code{TRUE} (not the default), the return value contains two extra
items, \code{foldid} and \code{foldFits}, that may be used to reassess the crossvalidation value}

\item{verbosity}{The higher this value, the more levels of progress and debug
information is displayed (note: in R for Windows, turn off buffered output)}
}
\value{
An object of class \code{cv.pim}, of class \code{cv.glmnet}, and depending on the type
	of linkfunction (see \code{cv.glmnet}) some more classes.
It holds all the items of the original \code{pimob} and all necessary items for a
	\code{cv.glmnet} object (so the relevant S3 methods like \code{print} and \code{plot} will
	work for them.)
}
\description{
Crossvalidate a glmnet PIM
}
\examples{
set.seed(1)
pen.N<-100
pen.noisep<-50
pen.noisemat<-matrix(rnorm(pen.N*pen.noisep), nrow=pen.noisep)
pendta<-data.frame(y=rnorm(pen.N), x=factor(sample(2, pen.N, replace=TRUE)), pen.noisemat)
pendta$y[pendta$x=="2"]<-pendta$y[pendta$x=="2"]+1
colnames(pendta)[(1:pen.noisep)+2]<-paste("X", 1:pen.noisep, "X", sep="")
pen.formula<-paste("y~", paste(c("F(x)", setdiff(colnames(pendta), c("x", "y"))), collapse="+"), sep="")

penpim<-pim(as.formula(pen.formula), data=pendta, link="identity", poset=noselfposet,
	estimator=estimator.glmnet(), varianceestimator=NULL, keep.data=TRUE, verbosity=0,
	interpretation="regular")
cv.penpim.naive<-crossvalidate.pim(penpim, method="naive")
cv.penpim<-crossvalidate.pim(penpim, method="fullsplit")
cv.penpim.ps<-crossvalidate.pim(penpim, method="semisplit")
}
\seealso{
\code{\link{pim}} \code{cv.glmnet}
}
\keyword{crossvalidate}
\keyword{pim}

