% Generated by roxygen2 (4.0.1): do not edit by hand
\name{demo.WilcoxonMannWhitney}
\alias{demo.BrownHettmansperger}
\alias{demo.JonckheereTerpstra}
\alias{demo.KruskalWallis}
\alias{demo.MackSkillings}
\alias{demo.MackWolfe}
\alias{demo.WilcoxonMannWhitney}
\title{Generate sample data that is appropriate for the classical distributionfree tests}
\usage{
demo.WilcoxonMannWhitney(n = 100, groupmeans = c(0, 1),
  force.balanced = FALSE)

demo.KruskalWallis(n = 100, groupmeans = c(0, 1, 2),
  force.balanced = FALSE)

demo.MackSkillings(blockeffects = rnorm(4), groupmeans = c(0, 1, 2),
  replications = 10, force.balanced = FALSE)

demo.JonckheereTerpstra(n = 100, groupmeans = c(0, 1, 2, 4),
  force.balanced = FALSE)

demo.MackWolfe(n = 100, groupmeans = c(0, 1, 5, 2),
  force.balanced = FALSE)

demo.BrownHettmansperger(n = 100, groupmeans = c(0, 1, 2),
  force.balanced = FALSE)
}
\arguments{
\item{n}{Number of observations}

\item{groupmeans}{Means of normal outcome distribution for each group (note this implies the number of groups)}

\item{force.balanced}{If \code{TRUE}, the design is forced to be balanced (an error occurs if this is impossible)}

\item{blockeffects}{Effect of each block (note this implies the number of blocks)}

\item{replications}{Number of replications per group/block combination}
}
\value{
a list holding the following items:
\item{dta}{A \code{data.frame} holding the generated data. The outcome is column \code{"y"},
	the predictor is \code{"x"} and the blocking variable (if relevant) is \code{"b"}}
\item{n}{As passed in.}
\item{groupmeans}{As passed in.}
For \code{demo.MackSkillings} these additional items are present:
\item{blocks}{Number of blocks.}
\item{blockeffects}{(Random) effect of each specific block.}
\item{treatments}{Number of groups in the predicting variable.}
\item{tb1}{\code{data.frame} holding all combinations of groups and blocks.}
\item{tbmult}{Similar as \code{tb1}, but now holding the repetitions and also randomized.}
}
\description{
Generate sample data that is appropriate for the classical distributionfree tests
}
\keyword{data}
\keyword{legacy}
\keyword{pim}

