% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EngelData.R
\docType{data}
\name{EngelData}
\alias{EngelData}
\title{Engel's food expenditure data}
\format{A data frame of 235 observations and 2 variables:
\itemize{
	\item{\code{foodexp}} Food expenditure (FE)
	\item{\code{income}} Household income (HI)
}}
\source{
Koenker, R. and Bassett, G (1982) Robust Tests of Heteroscedasticity
based on Regression Quantiles; \emph{Econometrica 50, 43-61}.
}
\usage{
EngelData
}
\description{
This data originates from Koenker and Bassett (1982). It is a regression
data set consisting of 235 observations on income and expenditure on
food for Belgian working class households. The dataset is copied from
the \href{https://cran.r-project.org/web/packages/quantreg/index.html}{quantreg} package.
}
\keyword{data}
\keyword{engel}
