% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEVData.R
\docType{data}
\name{FEVData}
\alias{FEVData}
\title{Forced expiratory volume in children}
\format{A data frame with 654 observations and the following 5 variables:
\itemize{
	\item{\code{FEV}} Forced Expiratory Volme (FE)
	\item{\code{Age}} Age of the child
	\item{\code{Height}} Height of the child
	\item{\code{Sex}} Gender of the child (1 for boys, 0 for girls)
	\item{\code{Smoke}} 1 if the child smokes, 0 otherwise
}}
\source{
Rosner(1999) : Fundamentals of Biostatistics. \emph{Pacific Grove: Duxbury}
}
\usage{
FEVData
}
\description{
This dataset is a classic example originating from a number of studies
done by Tager et al in the seventies and eighties. This data contains a
cross-sectional subset of the original data used for investigating
the relationship between forced expiratory volume (FEV) and smoking
status. The data is used as an example dataset in the book
Fundamentals of Biostatistics by Rosner (1999).
}
\keyword{FEV}
\keyword{data}
