% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.posfun.R
\name{.make.posfun}
\alias{.make.posfun}
\title{Create a poset function}
\usage{
.make.posfun(poset)
}
\arguments{
\item{poset}{a vector with the columns as indices}
}
\value{
A function that takes a single vector as argument, and that
returns the vector with the poset vector applied to it.
}
\description{
This function creates a poset function from a poset. The function is not
exported and shouldn't be called by the user.
}
