% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.pim.R
\name{new.pim}
\alias{new.pim}
\title{Create an object of class pim}
\usage{
new.pim(...)
}
\arguments{
\item{...}{Data to include in the new object. See \code{\link[methods]{new}}}
}
\value{
an object of class \code{pim}. See \code{\link{pim-class}}
}
\description{
This function is the constructor for an object of class pim. It is nothing
but a placeholder for \code{new("pim", ...)}. This function is not exported.
}
