% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oVisDrugAtt.r
\name{oVisDrugAtt}
\alias{oVisDrugAtt}
\title{Function to visualise enrichment results using a upset-like plot}
\usage{
oVisDrugAtt(
  data,
  member.sortBy = NULL,
  color = "cyan4",
  shape = 18,
  size = 2,
  drugs.customised = NULL
)
}
\arguments{
\item{data}{a data frame. It contains all these columns including  'value' and 'member' (separated by ',' or ', ')}

\item{member.sortBy}{a data frame. It contains all the target nodes and their weights for ranking.}

\item{color}{the point color}

\item{shape}{the point shape}

\item{size}{the point size}

\item{drugs.customised}{the customised shown of drugs in the combination matrix.}

\item{label.height.unit}{the unit specifying the per-row height of the combination matrix. By default, it is NULL. If specified (such as 8), it will be used to decide 'combmatrix.label.height' and 'combmatrix.label.text'}

\item{drug.levels}{how to define the levels of drug in the combination matrix It can be 'num' (the number of non-zeros for a member) or 'customised' (see 'levels.customised' below) or 'auto' (alphabeta order; by default)}
}
\value{
an object of class "ggplot"
}
\description{
\code{oVisDrugAtt} is supposed to visualise enrichment results using a upset-like plot. The top is the kite plot, and visualised below is the combination matrix for overlapped genes. It returns an object of class "ggplot".
}
\note{
none
}
\examples{
 \dontrun{
set.seed(85)
data <- rbind(df_best, df_single) \%>\% distinct() \%>\% 
    arrange(i,frac.disconnected) \%>\%
    select(drug,frac.disconnected, nodes.removed) \%>\%
    rename_with(~c("drug","value","member")) 
    p<- data \%>\% oVisDrugAtt()
    }
}
\seealso{
\code{\link{oVisDrugAtt}}
}
