\docType{package}
\name{pim-package}
\alias{pim-package}
\title{Probabilistic Index Models}
\description{
  Fit a probabilistic index model.
}
\details{
  \tabular{ll}{ Package: \tab pim\cr Type: \tab Package\cr
  Version: \tab 1.1.0\cr Date: \tab 2012-11-19\cr License:
  \tab GPL (>= 2)\cr LazyLoad: \tab yes\cr Depends: R (>=
  2.12), Matrix, glmnet, BB, nleqslv, MASS\cr}
}
\author{
  Jan De Neve \email{JanR.DeNeve@UGent.be}
}
\references{
  [PENDING]
}
\keyword{package}

