% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pim.poset-class.R
\docType{class}
\name{pim.poset-class}
\alias{pim.poset-class}
\alias{pim.poset}
\title{The pim.poset class}
\description{
The pim.poset class is an S4 class that inherits from \code{\link{environment}}
and contains the poset functions for a \code{\link{pim}}. It's a class
used internally and should not be adapted by the user. The correct
interpretation of the formula is dependent on this object. The object
mainly functions as a slot in object of class 
\code{\link{pim.environment}}.
}
\section{Slots}{

\describe{
\item{\code{compare}}{a character value with the type of poset. This can
take the values "unique", "all" and "custom".}

\item{\code{nobs}}{an integer value describing the number of observations 
for which this poset is meant to be used.}
}}

\section{Note}{

The pim.poset class doesn't really make sense to be used on itself.
It is part of the \code{\link{pim.environment}} class and shouldn't
be used outside this context.
}

