% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUData.R
\docType{data}
\name{SUData}
\alias{SUData}
\title{This is the Surgical unit data}
\description{
This is the Surgical unit data
}
\details{
These are the columns and their meanings
\itemize{
	\item{\code{EnT}} Enzyme function test score
	\item{\code{Gender}} Gender of the patient (0: male, 1: female)
	\item{\code{Alcohol}} History of alcohol use (0: none, 1: moderate, 2: severe)
	\item{\code{SurvivalTime}} Survival time of each patient (the outcome of interest)
}
}
\keyword{SU}
\keyword{data}
