% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{tidy.pim}
\alias{tidy.pim}
\alias{pim_tidiers}
\title{Tudy of a pim object}
\usage{
tidy.pim(x, conf.int = FALSE, conf.level = 0.95, expit = FALSE, ...)
}
\arguments{
\item{x}{An `pim` object created by [pim::pim()].}

\item{conf.int}{Logical indicating whether or not to include
a confidence interval in the tidied output. Defaults to FALSE.}

\item{conf.level}{The confidence level to use for the confidence
interval if conf.int = TRUE. Must be strictly greater than 0
and less than 1. Defaults to 0.95, which corresponds to a
95 percent confidence interval}

\item{expit}{Logical indicating whether or not to apply [plogis] to the
coefficient estimates. Defaults to FALSE.}

\item{...}{Unused, included for generic consistency only.}
}
\value{
A tidy [tibble::tibble()] summarizing component-level
  information about the model
}
\description{
Tudy of a pim object
}
\examples{
\dontshow{if (rlang::is_installed("ggplot2")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

library(ggplot2)
library(dplyr)
library(pim)
data('FEVData')

mod <- pim(FEV~ Smoke*Sex , data=FEVData)

tidy(mod)

# coefficient plot
d <- tidy(mod, conf.int = TRUE)

ggplot(d, aes(estimate, term, xmin = conf.low, xmax = conf.high, height = 0)) +
  geom_point() +
  geom_vline(xintercept = 0, lty = 4) +
  geom_errorbarh()

# probabilistic index plot
d <- tidy(mod, conf.int = TRUE, expit = TRUE)

ggplot(d, aes(estimate, term, xmin = conf.low, xmax = conf.high, height = 0)) +
  geom_point() +
  geom_vline(xintercept = 0, lty = 4) +
  geom_errorbarh()
\dontshow{\}) # examplesIf}
}
\seealso{
[tidy()], [pim::summary.pim()]
}
\concept{pim tidiers}
