% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.internal.R
\name{vcov.internal}
\alias{vcov.internal}
\alias{U.sandwich}
\alias{U.score}
\title{Internal functions for vcov estimation}
\usage{
U.sandwich(Zbeta, Z, Y, link, W = NULL)

U.score(Zbeta, Z, Y, link, W = NULL)
}
\arguments{
\item{Zbeta}{fitted values}

\item{Z}{design matrix}

\item{Y}{pseudo responses}

\item{link}{character vector with link function}

\item{W}{vector with weights}
}
\description{
These functions serve as preparation functions to calculute the variance-
covariance matrix of a pim using any of the \code{\link{vcov.estimators}}
provided in this package. The result of these preparation functions
is used by the \code{\link{sandwich.estimator}} and 
\code{\link{score.estimator}} functions respectively.
}
\note{
These functions should NOT be called by the user
}
