% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.matrix.pim.summary.R
\name{as.matrix.pim.summary}
\alias{as.matrix.pim.summary}
\alias{as.matrix}
\alias{as.matrix,pim.summary-method}
\alias{as.matrix,pim-method}
\title{Convert a pim.summary object to a matrix}
\usage{
as.matrix(x, ...)

\S4method{as.matrix}{pim.summary}(x, ...)

\S4method{as.matrix}{pim}(x, ...)
}
\arguments{
\item{x}{a \code{pim.summary} object}

\item{...}{additional arguments to be passed to or from methods.
This one is ignored.}
}
\value{
a matrix with the estimate, standard error, Z value and
probability for every coefficient.
}
\description{
This function converts a summary object to a matrix so values can be
extracted from it.
}
