% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.pim.R
\name{confint.pim}
\alias{confint.pim}
\alias{confint.pim.summary}
\alias{confint}
\alias{confint,pim-method}
\alias{confint,pim.summary-method}
\title{Calculate Wald confidence intervals around the coefficients of a PIM}
\usage{
confint(object, parm, level = 0.95, ...)

\S4method{confint}{pim}(object, parm, level = 0.95, ...)

\S4method{confint}{pim.summary}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{a \code{\link{pim}} or \code{\link{pim.summary}} object}

\item{parm}{a specification of which parameters are to be given confidence intervals. Either a vector of numbers or a vector of names.
If missing, all parameters are considered}

\item{level}{The confidence level required.}

\item{...}{extra arguments to methods}
}
\description{
This function returns Wald confidence intervals around the 
coefficients of a fitted \code{\link{pim}} object
}
