% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EngelData.R
\docType{data}
\name{EngelData}
\alias{EngelData}
\title{This is the engel data}
\description{
This is the engel data
}
\details{
These are the columns and their meanings
\itemize{
	\item{\code{foodexp}} Food expenditure (FE)
	\item{\code{income}} Household income (HI)
}
}
\keyword{data}
\keyword{engel}
