% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEVData.R
\docType{data}
\name{FEVData}
\alias{FEVData}
\title{This is the Childhood respiratory disease data}
\description{
This is the Childhood respiratory disease data
}
\details{
These are the columns and their meanings
\itemize{
	\item{\code{FEV}} Forced Expiratory Volme (FE)
	\item{\code{Age}} Age of the child
	\item{\code{Height}} Height of the child
	\item{\code{Sex}} Gender of the child (1 for boys, 0 for girls)
	\item{\code{Smoke}} 1 if the child smokes, 0 otherwise
}
}
\keyword{FEV}
\keyword{data}
