% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHData.R
\docType{data}
\name{MHData}
\alias{MHData}
\title{Mental Health in Alachua County}
\format{An object of class \code{data.frame} with 40 rows and 3 columns.}
\source{
Agresti, A. (2007) An Introduction to Categorical Data Analysis.
Hoboken: Wiley.
}
\usage{
MHData
}
\description{
This dataset originates from Agresti (2007) and consists of mental health
data for a random sample of 40 adult residents of Alachua County, Florida.
}
\details{
These are the columns and their meanings
\itemize{
	\item{\code{mental}}: Mental impairment (MI) as an ordinal variable with
	categories 1 (well), 2 (mild symptom formation), 3 (moderate symptom
	formation) and 4 (impaired)
	\item{\code{ses}}: Socioeconomic status as a binary variable, coded as
	0 (low socio-economic status) and 1 (high socio-economic status)
	\item{\code{life}}: Life index (LI) as a composite measure that
	quantifies the severity and number of important life events such as
	birth of a child, death in the family and divorce.
}
}
\keyword{MH}
\keyword{MentalHealth}
\keyword{data}
