% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHData.R
\docType{data}
\name{MHData}
\alias{MHData}
\title{This is the Mental health data}
\description{
This is the Mental health data
}
\details{
These are the columns and their meanings
\itemize{
	\item{\code{mental}} Mental impairment (MI)
	\item{\code{ses}} Socioeconomic status
	\item{\code{life}} Life index (LI)
}
}
\keyword{FEV}
\keyword{data}
