\docType{package}
\name{pim-package}
\alias{pim-package}
\title{Probabilistic Index Models}
\description{
  Fit a probabilistic index model.
}
\details{
  \tabular{ll}{ Package: \tab pim\cr Type: \tab Package\cr
  Version: \tab 1.1.1.1\cr Date: \tab 2013-03-20\cr
  License: \tab GPL (>= 2)\cr LazyLoad: \tab yes\cr
  Depends: R (>= 2.12), Matrix, glmnet (>= 1.9-3), BB,
  nleqslv, MASS, lqa\cr}
}
\author{
  Jan De Neve \email{JanR.DeNeve@UGent.be} Nick Sabbe
  \email{Nick.Sabbe@UGent.be}
}
\references{
  [PENDING]
}
\keyword{package}

